% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossBatchGrouping.R
\name{CrossBatchGrouping}
\alias{CrossBatchGrouping}
\title{Title Cross Batch ICA Source Component Grouping
Group the expression programs (source components) generated from different batch of scRNA-seq dataset through PAM clustering.}
\usage{
CrossBatchGrouping(
  ica.pooling,
  k.max = (ncol(ica.pooling) - 1),
  plot = TRUE,
  cor = "pearson",
  W.top = 2.5,
  filtering = TRUE,
  threshold = 30,
  Unique.Preservation = TRUE
)
}
\arguments{
\item{ica.pooling}{A matrix object that including all ICA source components generate from \emph{ICAcomputing},}

\item{k.max}{the maximum clusters number of ica-components (default: the number of ica-components - 1)}

\item{plot}{whether plot cross batch ica source component correlation map}

\item{cor}{the correlation method which determine the correlation between different pair of source components}

\item{W.top}{the threshold to determine the activated genes, the genes which has absolute attributes value large than threshold*standard derivation from mean are the activated genes (default: 2.5)}

\item{filtering}{whether filter the components which their 'activated gene' number lower than a specific values}

\item{threshold}{the threshold number to determine which component need to be filtered (default: 30)}

\item{Unique.Preservation}{whether to preserve the cluster with only one component included (default: TRUE)}
}
\value{
a list object which containing the filtered ica-components matrix.
}
\description{
Title Cross Batch ICA Source Component Grouping
Group the expression programs (source components) generated from different batch of scRNA-seq dataset through PAM clustering.
}
