% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleTool.R
\name{FindMarkerModule}
\alias{FindMarkerModule}
\title{Title Module markers of identity classes
Finds markers (differentially activated modules) for identity classes}
\usage{
FindMarkerModule(
  obj,
  identity,
  assay = "IcaNet",
  MQC = TRUE,
  pvalue = 0.05,
  rank.order = "myAUC"
)
}
\arguments{
\item{obj}{a Seurat object}

\item{identity}{the identity id}

\item{assay}{which assay shall be tested (Default: "IcaNet")}

\item{MQC}{whether to control the statistical significance level of module (Default: TRUE)}

\item{pvalue}{integer value to indicate the threshold to filter the module (Default: 0.05)}

\item{rank.order}{rank the results based on which feature (Default: "myAUC")}
}
\value{
a matrix object which containing the significant module information
}
\description{
Title Module markers of identity classes
Finds markers (differentially activated modules) for identity classes
}
