% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAcomputing.R
\name{ICAcomputing}
\alias{ICAcomputing}
\title{Title Using Independent Component Analysis to decompose scRNA-seq gene expression matrix}
\usage{
ICAcomputing(
  obj,
  center = TRUE,
  scale = FALSE,
  RMT = TRUE,
  nc.vec = NULL,
  nc.global.mode = NULL,
  global.mode = FALSE,
  ICA.type = "JADE",
  nbIt = 10,
  funClus = "hclust",
  two.stage = TRUE,
  two.stage.ICA = "JADE"
)
}
\arguments{
\item{obj}{a seurat object which contain the gene expression matrix of all batches, meanwhile it also contain the \emph{batch} ID which document the batch label information}

\item{center}{a boolean variable to determine whether perform centralization to each batch}

\item{scale}{a boolean variable to determine whether perform scaling to each batch}

\item{RMT}{a boolean variable to determine whether perform random matrix theory based number of components estimation}

\item{nc.vec}{a numrical vector to determine the number of independent components for each batch}

\item{nc.global.mode}{a numeric vector to determine the number of components of all batches when running global mode.}

\item{global.mode}{a boolean variable to determine whether running global mode based ICA}

\item{ICA.type}{an optional character string given a method for computing independent components. One of "JADE" (default), "FastICA", or "MineICA": can be abbreviated}

\item{nbIt}{The number of iterations of FastICA in MineICA}

\item{funClus}{The clustering function to be used to cluster the estimates in MineICA. One of "hclust"(default),"kmeans", "pam", "agnes" can be abbreviated}

\item{two.stage}{Whether running ICA+RMT based components number estimation}

\item{two.stage.ICA}{Which type of ICA methods shall be applied in ICA+RMT estimate. One can select "FastICA" or "JADE"}
}
\value{
a list object which containing the ica-components
}
\description{
We introduced independent component analysis (ICA) into single cell clustering to decompose the gene expression matrix into a number of independent components. Each independent component was characterized by a co-expression pattern and was associated with certain meaningful biological pathway.
}
