% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleSignificanceTest.R
\name{ModuleSignificanceTest}
\alias{ModuleSignificanceTest}
\title{Title Module Significance Test}
\usage{
ModuleSignificanceTest(Data, statl.v, geneSets, nMC, verbose = FALSE)
}
\arguments{
\item{Data}{a Z-score transformed matrix}

\item{statl.v}{the ica-components}

\item{geneSets}{the module sets which need to be tested its significance}

\item{nMC}{the number of permutations}

\item{verbose}{a boolean variable, whether show the running process (default: FALSE)}
}
\value{
a integer value which indicating the significance of modules
}
\description{
Title Module Significance Test
}
