% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleTool.R
\name{plot_module}
\alias{plot_module}
\title{Title Plot interactive molecular network}
\usage{
plot_module(
  gene_sets,
  network,
  ica.score,
  nodeSize = "ica.score",
  fontSize = 15,
  charge = -100,
  width = 500,
  height = 500,
  size_scale = 1
)
}
\arguments{
\item{gene_sets}{the module needed to plot the interaction network}

\item{network}{the molecular network}

\item{ica.score}{the ica-components}

\item{nodeSize}{the character value indicating node size is showed according to which information ("ica.score" or "degree")}

\item{fontSize}{the integer value indicating the size of font (Default: 15)}

\item{charge}{the integer value indicating the degree of expanding of network (Default: -100)}

\item{width}{width of the graph}

\item{height}{height of the graph}

\item{size_scale}{the size of the node (Default: 1)}
}
\value{
a interactive plot of molecular network
}
\description{
Title Plot interactive molecular network
}
