% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunICAnet.R
\name{RunICAnet}
\alias{RunICAnet}
\title{Title Using independent components to construct weighted molecular network to detect function module for single cell clustering
ICAnet used independent components to construct weighted PPI, and running walk-trap algorthm to detect module on it. The resulted modules are used for the following analysis}
\usage{
RunICAnet(
  obj,
  ica.filter,
  W.top = 2.5,
  PPI.net = NULL,
  species = 9606,
  score = 600,
  max.step = 10,
  small.size = 3,
  nMC = 100,
  aucMaxRank = 3000,
  cores = 6,
  ModuleSignificance = TRUE,
  scale = TRUE
)
}
\arguments{
\item{obj}{a Seurat object}

\item{ica.filter}{the filtered/unfiltered ica-components set}

\item{W.top}{the threshold to determine the activated genes, the genes which has absolute attributes value large than threshold*standard derivation from mean are the activated genes (default: 2.5)}

\item{PPI.net}{a matrix object which indicating the PPI network, a boolean network is required}

\item{max.step}{Integer number. The maximum step run in the walk-trapped based community detect.}

\item{small.size}{integer number to determine the minimum size of module. The module which has the number of gene member less than this value will be filtered}

\item{nMC}{the number of permutations, which is used for calculate the pvalue of each module (default: 100)}

\item{aucMaxRank}{Integer number. The number of highly-expressed genes to include when computing AUCell}

\item{cores}{the number of cores used for computation}

\item{ModuleSignificance}{the boolean variable to indicate whether perform module significant test (default: FALSE)}

\item{scale}{the boolean variable to indicate whether perform scaling over each batch of scRNA-seq gene expression data}
}
\value{
a Seurat object which contain the "IcaNet" assay
}
\description{
Title Using independent components to construct weighted molecular network to detect function module for single cell clustering
ICAnet used independent components to construct weighted PPI, and running walk-trap algorthm to detect module on it. The resulted modules are used for the following analysis
}
