% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModuleTool.R
\name{RunModuleSVD}
\alias{RunModuleSVD}
\title{Using SVD to decompose the module activity matrix}
\usage{
RunModuleSVD(obj, nu = 30, power = 0.5)
}
\arguments{
\item{obj}{a Seurat object}

\item{nu}{the number of singular vector}

\item{power}{integer value indicating the smoothness of signal on singular vector}
}
\value{
a Seurat object which containing reduction object "Module_SVD"
}
\description{
Using SVD to decompose the module activity matrix
}
