% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedBasalComponents.R
\name{SharedBasalComponents}
\alias{SharedBasalComponents}
\title{Identify shared basal components}
\usage{
SharedBasalComponents(
  data_list,
  var.names,
  reduct.dim,
  center = TRUE,
  method = "irlba",
  fastICA = FALSE,
  corr = "spearman",
  W.top = 2.5
)
}
\arguments{
\item{data_list}{a list Seurat objects}

\item{var.names}{the features genes, e.g highly variable features learned from \emph{FindVariableFeatures}}

\item{reduct.dim}{the reducted dimensions, default: 30}

\item{method}{truncted SVD solver. "irlba" or "rARPACK"}

\item{fastICA}{boolean variable,if TRUE, using FastICA, otherwise use JADE to perform ICA.}

\item{corr}{the correlation method which determine the correlation between different pair of source components (default: spearman)}

\item{W.top}{the threshold to determine the activated genes, the genes which has absolute attributes value large than threshold*standard derivation from mean are the activated genes (default: 2.5)}
}
\value{
a list object which containing merged seurat object and basal components
}
\description{
Identify shared basal components through using whiten matrices derived from latent semantic indexing
}
\examples{
\dontrun{
list <- SharedBasalComponents(PBMC_list,features,30)
objs <- list$mergedObject
basalcomp <- list$basalcomp
}
}
