% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICAnetTF.R
\name{TF_Net_Generate}
\alias{TF_Net_Generate}
\title{Convert feather file into motif-gene boolean network}
\usage{
TF_Net_Generate(feather.readPath, cutoff = 1)
}
\arguments{
\item{feather.readPath}{a character value to indicate the path which saved the feather file}

\item{cutoff}{a cutoff to determine how to filter the unwanted Motif-gene links (default: 1)}
}
\value{
a matrix object which containing the boolean network of TF binding motif-gene
}
\description{
This function is provided for convert feather file into motif-gene boolean network, the feather file containing the motif binding informtion for human/mouse/fly species, the files could be downloaded from cisTarget databases https://resources.aertslab.org/cistarget/
}
