% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstNumModule.R
\name{EstNumModule}
\alias{EstNumModule}
\title{Title Estimate the number of components
an Random Matrix theory based component number estimation}
\usage{
EstNumModule(data.m, default = TRUE, svd.num = NULL)
}
\arguments{
\item{data.m}{a matrix object which need to be estimated its number of components}

\item{default}{boolean variable to determine whether restrict the maximum number of decomposed components}

\item{svd.num}{an integer value to indicate maximum number of decomposed components}
}
\value{
a integer value which indicating the estimated number of ica components
}
\description{
Title Estimate the number of components
an Random Matrix theory based component number estimation
}
